Added PSResourceHelper class to share methods between Publish and Compress
