+++ src/code/CompressPSResource.cs
@@ -20,8 +20,6 @@ namespace Microsoft.PowerShell.PSResourceGet.Cmdlets
     {
         #region Parameters
 
-        private PublishPSResource _publishPSResource;
-
         /// <summary>
         /// Specifies the path to the resource that you want to compress. This parameter accepts the path to the folder that contains the resource.
         /// Specifies a path to one or more locations. Wildcards are permitted. The default location is the current directory (.).
@@ -40,6 +38,12 @@ namespace Microsoft.PowerShell.PSResourceGet.Cmdlets
 
         #endregion
 
+        #region Members
+
+        private PSResourceHelper _psResourceHelper;
+
+        #endregion
+
         #region Method Overrides
 
         protected override void BeginProcessing()
@@ -48,78 +52,17 @@ namespace Microsoft.PowerShell.PSResourceGet.Cmdlets
             // This is to create a better experience for those who have just installed v3 and want to get up and running quickly
             RepositorySettings.CheckRepositoryStore();
 
-            string resolvedPath = null;
-            string resolvedDestinationPath = null;
+            _psResourceHelper = new PSResourceHelper(
+                this,
+                Path,
+                DestinationPath);
 
-            try
-            {
-                resolvedPath = GetResolvedProviderPathFromPSPath(Path, out ProviderInfo provider).First();
-            }
-            catch (MethodInvocationException)
-            {
-                // path does not exist
-                ThrowTerminatingError(new ErrorRecord(
-                    new ArgumentException("The path to the resource to publish does not exist, point to an existing path or file of the module or script to publish."),
-                    "SourcePathDoesNotExist",
-                    ErrorCategory.InvalidArgument,
-                    this));
-            }
-
-            if (!String.IsNullOrEmpty(DestinationPath))
-            {
-                resolvedDestinationPath = GetResolvedProviderPathFromPSPath(DestinationPath, out ProviderInfo provider).First();
-
-                if (Directory.Exists(resolvedDestinationPath))
-                {
-                    DestinationPath = resolvedDestinationPath;
-                }
-                else
-                {
-                    try
-                    {
-                        Directory.CreateDirectory(resolvedDestinationPath);
-                    }
-                    catch (Exception e)
-                    {
-                        ThrowTerminatingError(new ErrorRecord(
-                            new ArgumentException($"Destination path does not exist and cannot be created: {e.Message}"),
-                            "InvalidDestinationPath",
-                            ErrorCategory.InvalidArgument,
-                            this));
-                    }
-                }
-            }
-
-            _publishPSResource = new PublishPSResource(true, Path, resolvedDestinationPath, resolvedPath);
-
-            try
-            {
-                _publishPSResource.CheckAllParameterPaths();
-            }
-            catch (Exception e)
-            {
-                WriteError(new ErrorRecord(
-                    new ArgumentException(e.Message),
-                    "InvalidPath",
-                    ErrorCategory.InvalidArgument,
-                    this));
-            }
+            _psResourceHelper.CheckAllParameterPaths();
         }
 
         protected override void EndProcessing()
         {
-            try
-            {
-                _publishPSResource.PackAndPush();
-            }
-            catch (Exception e)
-            {
-                WriteError(new ErrorRecord(
-                    new ArgumentException(e.Message),
-                    "InvalidPath",
-                    ErrorCategory.InvalidArgument,
-                    this));
-            }
+            _psResourceHelper.PackResource();
         }
 
         #endregion
